+++ drizzle/0001_glorious_venom.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "pending_session" DROP CONSTRAINT "pending_session_email_unique";